LayoutTests/imported/w3c/web-platform-tests/reporting/document-reporting-not-batch-different-document.https-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Refused to load https://localhost:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.


FAIL Reports are not batched for same url in different document. assert_equals: Number of reports expected 2 but got 0
